We are using fast-sass-loader (rather than sass-loader) but that lacked a way to pass the "compressed" outputStyle option to node-sass. I opened an issue two days ago with yibn2008/fast-sass-loader, but it hasn't even been looked at since, so I initially implemented it in a fork of fast-sass-loader. Doing that makes the size of our webpacked .js files much smaller because it removes the comments from the output css. However, rush does not cooperate with loading from an npm fork, so I had to abandon that option.

I also changed to using mini-css-extract-loader in production builds, so that the css files are separate from the javascript files after webpacking. That required changes to IModelJsLoader to load those files, and to buildIModelJs to put the output .css files into the correct lib/webserver directories. I added the optimize-css-assets-webpack-plugin to get the css files minimized for production (so the change to fast-sass-loader isn't needed for production).

The development modules (which contain the css since they don't run mini-css-extract-loader) are still too large because the css is not compressed. Hopefully the issue with fast-sass-loader will be addressed soon.

+++ tools/webpack/modules/buildIModelJsModule.ts
@@ -124,10 +124,12 @@ class Utils {
     }
   }
 
-  public static moveFile(sourceDirectory: string, destDirectory: string, fileName: string) {
+  public static moveFile(sourceDirectory: string, destDirectory: string, fileName: string, warn: boolean) {
     const sourceFile = path.join(sourceDirectory, fileName);
     if (!fs.existsSync(sourceFile)) {
-      console.log("Error: Trying to move file that does not exist", sourceFile);
+      if (warn)
+        console.log("Error: Trying to move file that does not exist", sourceFile);
+      return;
     }
     const destFile = path.join(destDirectory, fileName);
     fs.renameSync(sourceFile, destFile);
@@ -229,6 +231,11 @@ class ModuleCopier {
 
   // symlinks the module file and source map file if available.
   private symlinkOrCopyModuleFile(moduleSourceFile: string, outFilePath: string) {
+    const mapFile = moduleSourceFile + ".map";
+    const outMapFile = outFilePath + ".map";
+    const cssFile = moduleSourceFile.replace(".js", ".css");
+    const outCssFile = outFilePath.replace(".js", ".css");
+
     if (this._alwaysCopy) {
       // copy  the module file.
       if (this._detail > 3)
@@ -240,11 +247,7 @@ class ModuleCopier {
 
       fs.copyFileSync(moduleSourceFile, outFilePath);
 
-      // if there's a source map file, link that, too.
-      const mapFile = moduleSourceFile + ".map";
-
       if (fs.existsSync(mapFile)) {
-        const outMapFile = outFilePath + ".map";
         if (this._detail > 3)
           console.log("  Copying file", mapFile, "to", outMapFile);
 
@@ -252,17 +255,26 @@ class ModuleCopier {
         if (fs.existsSync(outMapFile))
           fs.unlinkSync(outMapFile)
 
-        fs.copyFileSync(mapFile, outMapFile);
+        if (fs.existsSync(mapFile))
+          fs.copyFileSync(mapFile, outMapFile);
+
+        if (fs.existsSync(outCssFile))
+          fs.unlinkSync(outCssFile)
+
+        if (fs.existsSync(cssFile))
+          fs.copyFileSync(cssFile, outCssFile);
+
       }
     } else {
       // symlink the module file.
       this.ensureSymlink(moduleSourceFile, outFilePath);
 
       // if there's a source map file, link that, too.
-      const mapFile = moduleSourceFile + ".map";
-      const outMapFile = outFilePath + ".map";
       if (fs.existsSync(mapFile))
         this.ensureSymlink(mapFile, outMapFile);
+
+      if (fs.existsSync(cssFile))
+        this.ensureSymlink(cssFile, outCssFile);
     }
   }
 
@@ -775,10 +787,11 @@ class IModelJsModuleBuilder {
             try {
               const destPath = path.resolve(outputPath, "v" + version);
               Utils.makeDirectoryNoError(destPath);
-              Utils.moveFile(outputPath, destPath, "main.js");
-              Utils.moveFile(outputPath, destPath, "main.js.map");
-              Utils.moveFile(outputPath, destPath, "runtime.js");
-              Utils.moveFile(outputPath, destPath, "runtime.js.map");
+              Utils.moveFile(outputPath, destPath, "main.js", true);
+              Utils.moveFile(outputPath, destPath, "main.js.map", true);
+              Utils.moveFile(outputPath, destPath, "main.css", false);
+              Utils.moveFile(outputPath, destPath, "runtime.js", true);
+              Utils.moveFile(outputPath, destPath, "runtime.js.map", true);
             } catch (moveError) {
               resolve(new Result(operation.concat(" (move file)"), 1, moveError));
             }
